refactor: public solve with folders
