+++ frontend/src/lib/services/$public/sns.services.ts
@@ -1,0 +1,122 @@
+import {
+  queryAllSnsMetadata,
+  querySnsSwapCommitments,
+  querySnsSwapStates,
+} from "$lib/api/sns.api";
+import { loadProposalsByTopic } from "$lib/services/proposals.services";
+import { queryAndUpdate } from "$lib/services/utils.services";
+import {
+  snsProposalsStore,
+  snsQueryStore,
+  snsSwapCommitmentsStore,
+} from "$lib/stores/sns.store";
+import { toastsError } from "$lib/stores/toasts.store";
+import type { SnsSwapCommitment } from "$lib/types/sns";
+import type { QuerySnsMetadata, QuerySnsSwapState } from "$lib/types/sns.query";
+import { toToastError } from "$lib/utils/error.utils";
+import { Topic, type ProposalInfo } from "@dfinity/nns";
+
+export const loadSnsSummaries = (): Promise<void> => {
+  snsQueryStore.setLoadingState();
+
+  return queryAndUpdate<[QuerySnsMetadata[], QuerySnsSwapState[]], unknown>({
+    anonymousIdentity: true,
+    request: ({ certified, identity }) =>
+      Promise.all([
+        queryAllSnsMetadata({ certified, identity }),
+        querySnsSwapStates({ certified, identity }),
+      ]),
+    onLoad: ({ response }) => snsQueryStore.setData(response),
+    onError: ({ error: err, certified }) => {
+      console.error(err);
+
+      if (certified !== true) {
+        return;
+      }
+
+      // hide unproven data
+      snsQueryStore.setLoadingState();
+
+      toastsError(
+        toToastError({
+          err,
+          fallbackErrorLabelKey: "error__sns.list_summaries",
+        })
+      );
+    },
+    logMessage: "Syncing Sns summaries",
+  });
+};
+
+export const loadSnsSwapCommitments = (): Promise<void> => {
+  snsSwapCommitmentsStore.setLoadingState();
+
+  return queryAndUpdate<SnsSwapCommitment[], unknown>({
+    anonymousIdentity: true,
+    request: ({ certified, identity }) =>
+      querySnsSwapCommitments({ certified, identity }),
+    onLoad: ({ response: swapCommitments, certified }) => {
+      for (const swapCommitment of swapCommitments) {
+        snsSwapCommitmentsStore.setSwapCommitment({
+          swapCommitment,
+          certified,
+        });
+      }
+    },
+    onError: ({ error: err, certified }) => {
+      console.error(err);
+
+      if (certified !== true) {
+        return;
+      }
+
+      // hide unproven data
+      snsSwapCommitmentsStore.setLoadingState();
+
+      toastsError(
+        toToastError({
+          err,
+          fallbackErrorLabelKey: "error__sns.list_swap_commitments",
+        })
+      );
+    },
+    logMessage: "Syncing Sns swap commitments",
+  });
+};
+
+export const listSnsProposals = async (): Promise<void> => {
+  snsProposalsStore.setLoadingState();
+
+  return queryAndUpdate<ProposalInfo[], unknown>({
+    anonymousIdentity: true,
+    request: ({ certified, identity }) =>
+      loadProposalsByTopic({
+        certified,
+        identity,
+        topic: Topic.SnsAndCommunityFund,
+      }),
+    onLoad: ({ response: proposals, certified }) =>
+      snsProposalsStore.setProposals({
+        proposals,
+        certified,
+      }),
+    onError: ({ error: err, certified }) => {
+      console.error(err);
+
+      if (certified !== true) {
+        return;
+      }
+
+      // hide unproven data
+      snsProposalsStore.setLoadingState();
+
+      toastsError(
+        toToastError({
+          err,
+          fallbackErrorLabelKey: "error.proposal_not_found",
+        })
+      );
+    },
+    logMessage: "Syncing Sns proposals",
+  });
+};
