+++ frontend/src/tests/lib/components/launchpad/Proposals.spec.ts
@@ -3,7 +3,7 @@
  */
 
 import Proposals from "$lib/components/launchpad/Proposals.svelte";
-import { p_listSnsProposals } from "$lib/services/sns.services";
+import { listSnsProposals } from "$lib/services/$public/sns.services";
 import { snsProposalsStore } from "$lib/stores/sns.store";
 import { ProposalStatus, type ProposalInfo } from "@dfinity/nns";
 import { render, waitFor } from "@testing-library/svelte";
@@ -29,7 +29,7 @@ describe("Proposals", () => {
   it("should trigger listSnsProposals", () => {
     render(Proposals);
 
-    expect(p_listSnsProposals).toBeCalled();
+    expect(listSnsProposals).toBeCalled();
   });
 
   it("should not trigger listSnsProposals if already loaded", () => {
@@ -40,7 +40,7 @@ describe("Proposals", () => {
 
     render(Proposals);
 
-    expect(p_listSnsProposals).toBeCalled();
+    expect(listSnsProposals).toBeCalled();
   });
 
   it("should display skeletons", async () => {
